+++ operator/pkg/resources/target/statefulset.go
@@ -138,7 +138,11 @@ func NewInitContainerEnv(ais *aisv1.AIStore) (initEnv []corev1.EnvVar) {
 	initEnv = cmn.CommonInitEnv(ais)
 	initEnv = append(initEnv, cmn.EnvFromValue(cmn.EnvServiceName, headlessSVCName(ais.Name)))
 	if ais.Spec.TargetSpec.HostPort != nil {
-		initEnv = append(initEnv, cmn.EnvFromFieldPath(cmn.EnvPublicHostname, "status.hostIP"))
+		if ais.EnableNodeNameHost() {
+			initEnv = append(initEnv, cmn.EnvFromFieldPath(cmn.EnvPublicHostname, "spec.nodeName"))
+		} else {
+			initEnv = append(initEnv, cmn.EnvFromFieldPath(cmn.EnvPublicHostname, "status.hostIP"))
+		}
 	}
 	if ais.UseHostNetwork() {
 		initEnv = append(initEnv, cmn.EnvFromValue(cmn.EnvHostNetwork, "true"))
@@ -148,9 +152,6 @@ func NewInitContainerEnv(ais *aisv1.AIStore) (initEnv []corev1.EnvVar) {
 
 func NewAISContainerEnv(ais *aisv1.AIStore) []corev1.EnvVar {
 	baseEnv := cmn.CommonEnv()
-	if ais.Spec.TargetSpec.HostPort != nil {
-		baseEnv = append(baseEnv, cmn.EnvFromFieldPath(cmn.EnvPublicHostname, "status.hostIP"))
-	}
 	if ais.Spec.HasGCPBackend() {
 		baseEnv = append(baseEnv, cmn.EnvFromValue(cmn.EnvGoogleCreds, filepath.Join(cmn.DefaultGCPDir, "gcp.json")))
 	}